requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
cachetools==4.2.4
certifi==2024.8.30
charset-normalizer==3.3.2
exceptiongroup==1.2.2
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.9.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.62.3
grpcio-tools==1.66.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.7
hyperframe==6.0.1
idna==3.9
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
langchain-text-splitters==0.3.0
langsmith==0.1.120
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pillow==10.4.0
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.4
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.10.0
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.11.1
